Disable Pylint warning of camelCase names for Qt
Pylint warns on functions such as `mousePressEvent()`, `mouseMoveEvent()`,
`mouseReleaseEvent()`, `paintEvent()`, and `closeEvent()` because they're not
snake_case format but rather camelCase format, but we can't change that, as
those methods are part of the Qt API and we need to match their naming
convention.

Add local comments on those lines for Pylint to ignore these function names
only, while enabling checking the rest of the code.