Move check for `poppler-utils` before GUI starts
* Create a new `main()` function to encapsulate code; move startup code from
  top-level into the function
* Check for the presence of the Poppler utils we are using (`pdftocairo`,
  `pdftoppm`, `pdfinfo`) in the new `main()` function before starting the GUI
  rather than on GUI initialization

This means that the error wil be visible much sooner on the CLI output, as the
next step would be to run a package installation command anyway, and it's
easier to copy-paste the package name from the terminal.

Similarly, we are now providing a URL to click on for more installation
instructions, and that's easier to copy-paste or click on from the terminal
output rather from the GUI window (which is not copy-pastable or clickable,
without extra work, which we don't need to invest there).