Add a prototype of a web app with some caveats
The app currently doesn't work due to SVG export issues in PDF.js; see README.md
for details.

This prototype is part of https://github.com/mbrukman/pdf-extract-svg/issues/31
but it is not a usable app at this point. THE README.md provides some ideas and
suggestions for next steps, if anyone is interested in pursuing this further.